Model B: Invasion: add experiment log
I performed a bunch of experiments where a single noncarrier invades a
carrier-only population. I noticed some interesting phenomena, which led to
a bunch of followup experiments, documented in `INVASION.md`.

All parameters are logged in this file, along with observations. I ran the
experiments using the `make_data` script.

Signed-off-by: Zoltan A. Kocsis <11808286+zaklogician@users.noreply.github.com>