data: Fix division-by-zero in carrier-noncarrier ratio
The ratio of carriers to noncarriers appears in the data line. Previously, when
the number of noncarriers reached zero, this caused a division by zero, and
since the ratio is a floating-point value, `Infinity` was returned.

The presence of `Infinity` caused issues downstream, e.g. when plotting data,
since not all plotting tools can parse `inf` as a value.

To fix the issue, we now return the total number of cells in the world instead
of `Infinity` whenever the number of noncarriers is zero. This number exceeds
the number of carriers, and a fortiori all achievable carrier-noncarrier ratios
as well, so constitutes an effective standin for infinity.

Signed-off-by: Zoltan A. Kocsis <11808286+zaklogician@users.noreply.github.com>